+++ src/main/java/cn/v5cn/basicframework/dao/SystemResDao.java
@@ -10,9 +10,9 @@ import java.util.List;
  */
 @Repository
 public interface SystemResDao {
-    Long addSystemRes(SystemRes systemRes);
-    Long updateSystemRes(SystemRes systemRes);
-    Long deleteSystemRes(Long resId);
+    int addSystemRes(SystemRes systemRes);
+    int updateSystemRes(SystemRes systemRes);
+    int deleteSystemRes(Long resId);
     List<SystemRes> findAll();
     SystemRes findById(Long resId);
     int findByPidCount(Long pid);
